feat: added new images and updated index.tsx file
This commit adds two new images: dispatch.png and editor.png to the public/img directory. It also includes significant updates to the src/pages/index.tsx file, introducing new components, such as Navbar and Footer, and modifying the structure and content of the main page. We also provide new links to the vatACARS plugin source code and hub.