refactor: update how tokens are handled in login and oauth
This commit changes the way tokens are passed and handled in the login and oauth routes. Instead of using the request body, they are now passed as query parameters. This also includes checks for the existence of the tokens before proceeding. In addition, it changes the way tokens are stored in the user object, storing the first value of the array instead of the whole array. Finally, it removes the POST request to the login route in the oauth route, directly redirecting to the login route with the tokens as query parameters.