feat: add user authorization check and redirect in discord.ts
In this commit, an additional check has been added to verify if a user is authorized. If the user is not authorized, they are redirected to the "/api/oauth" path. Additionally, the user's nickname has been updated to include their first name and CID. These changes are found in the discord.ts file.