feat: added token invalidation scheduling
A new feature has been added to the API key generation functionality. This feature includes scheduling the invalidation of authentication tokens. This will happen automatically one month after the token is created. The addition of this feature enhances the security of our application by ensuring that tokens are not left valid indefinitely.