feat: add getStats API endpoint and update stats page
This commit introduces a new API endpoint /api/misc/getStats that fetches stats data from an external API. It also updates the stats page to use this new endpoint and processes the response data for display. The display of the "Authority", "Controlling Sectors", and "Frequency" information has been improved as well.