refactor: Update data fetching and style in Map component
This commit includes changes to the MapComponent. The data fetching has been changed to use 'network' instead of 'simawareStats' for more accurate data. Moreover, the text style of several elements in the map has been updated to use 'Montserrat' font for consistency and better readability. Also, the image source has been fixed to a single image, removing the random selection of images.